Refactor loading in ReferenceArrayInput
